LPD-55069 Prevent errors by not deleting the child's site pages, as they will be deleted on parent deletion

+++ modules/apps/headless/headless-admin-site/headless-admin-site-test/src/testIntegration/java/com/liferay/headless/admin/site/resource/v1_0/test/SitePageResourceTest.java
@@ -1045,6 +1045,12 @@ public class SitePageResourceTest extends BaseSitePageResourceTestCase {
 				Pagination.of(0, 0), null);
 
 		for (SitePage sitePage : page.getItems()) {
+			if (Validator.isNotNull(
+					sitePage.getParentSitePageExternalReferenceCode())) {
+
+				continue;
+			}
+
 			sitePageResource.deleteSiteSiteByExternalReferenceCodeSitePage(
 				testGroup.getExternalReferenceCode(),
 				sitePage.getExternalReferenceCode());
